pyproject.toml
[project]
name = "ms-fabric-mcp-server"
version = "0.6.0"
description = "Model Context Protocol (MCP) server for Microsoft Fabric - exposes Fabric operations as MCP tools for AI agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Microsoft" }
]
keywords = ["mcp", "fastmcp", "microsoft-fabric", "fabric", "ai", "agent", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastmcp>=2.10.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "azure-identity>=1.23.0",
    "requests>=2.25.0",
]

[project.optional-dependencies]
sql = [
    "pyodbc>=5.0.0",
]
telemetry = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-instrumentation-dbapi>=0.46b0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "responses>=0.23.0",
    "black>=23.0.0",
    "isort>=5.13.0",
    "bandit[toml]>=1.7.5",
]

[project.scripts]
ms-fabric-mcp-server = "ms_fabric_mcp_server.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ms_fabric_mcp_server"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=src/ms_fabric_mcp_server",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests requiring live services",
    "slow: Slow running tests",
]

[tool.bandit]
exclude_dirs = ["tests", "examples"]
skips = ["B101"]

[tool.coverage.run]
source = ["src/ms_fabric_mcp_server"]
